boot: Update fit_image_get_data() to use abuf
This function uses separate arguments for data and size. Use the new
abuf instead, so that they are paired and in one place. In some cases it
also saves an argument, thus potentially reducing code size.

Move the prototype to the header file while here.

+++ cmd/fpga.c
@@ -234,7 +234,7 @@ static int do_fpga_loadbp(struct cmd_tbl *cmdtp, int flag, int argc,
 static int do_fpga_loadmk(struct cmd_tbl *cmdtp, int flag, int argc,
 			  char *const argv[])
 {
-	size_t data_size = 0;
+	ulong data_size = 0;
 	void *fpga_data = NULL;
 #if defined(CONFIG_FIT)
 	const char *fit_uname = NULL;
